# spec/openapi.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Validation failure (invalid name, role, document, or malformed body)"
      },
      "Conflict": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Namespace already exists"
      },
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Valid token but insufficient role (caller can read but not write)"
      },
      "InternalServerError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Unexpected server error"
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Namespace missing **or** caller lacks read role. Both cases use\n404 so namespace existence is not leaked.\n"
      },
      "PayloadTooLarge": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Document or request body exceeds size limit"
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Missing or invalid Bearer token"
      }
    },
    "schemas": {
      "ACLUpdateRequest": {
        "properties": {
          "read_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "write_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "read_role",
          "write_role"
        ],
        "type": "object"
      },
      "CreateNamespaceRequest": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/NamespaceDocument",
            "description": "Initial document. Defaults to `{}` if omitted."
          },
          "name": {
            "description": "Lowercase alphanumeric, hyphens, and underscores. Max 64 chars.",
            "pattern": "^[a-z0-9_-]{1,64}$",
            "type": "string"
          },
          "read_role": {
            "description": "Role required to read the namespace. Must be `admin` or `user`.\n`write_role` must not be less restrictive than `read_role`\n(i.e. `write_role=user` requires `read_role=user`).\n",
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "write_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "read_role",
          "write_role"
        ],
        "type": "object"
      },
      "Error": {
        "properties": {
          "error": {
            "description": "Machine-readable code",
            "example": "not_found",
            "type": "string"
          },
          "message": {
            "description": "Human-readable description",
            "example": "namespace not found",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
      "NamespaceDocument": {
        "additionalProperties": true,
        "description": "Arbitrary JSON object stored verbatim by the server. Max 64 KB\nafter compaction. Arrays, scalars, and null are rejected.\n",
        "example": {
          "humidity": "home/sensors/humidity",
          "temperature": "home/sensors/temp"
        },
        "type": "object"
      },
      "NamespaceInfo": {
        "description": "Namespace identity and ACL roles, without document or timestamps.",
        "properties": {
          "name": {
            "example": "mqtt_topics",
            "pattern": "^[a-z0-9_-]{1,64}$",
            "type": "string"
          },
          "read_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "write_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "read_role",
          "write_role"
        ],
        "type": "object"
      },
      "NamespaceSummary": {
        "description": "Namespace identity, ACL roles, and timestamps. Returned by list.",
        "properties": {
          "created_at": {
            "example": "2024-01-15T09:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "example": "mqtt_topics",
            "pattern": "^[a-z0-9_-]{1,64}$",
            "type": "string"
          },
          "read_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "updated_at": {
            "example": "2024-01-15T09:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "write_role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "read_role",
          "write_role",
          "updated_at",
          "created_at"
        ],
        "type": "object"
      },
      "PutResult": {
        "properties": {
          "changed": {
            "description": "False when the submitted document is byte-identical (after\nJSON compaction) to the stored document. No write occurs\nand no backup is triggered in that case.\n",
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "changed"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Identity-issued user access token (service tokens are rejected).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Self-hosted structured-config service. Namespaces\nare JSON documents with per-namespace read/write role ACLs. Tokens\nare verified against the identity service's JWKS endpoint.\n\n## Role model\nEach namespace declares `read_role` and `write_role` (each `admin`\nor `user`). Admins always satisfy any role. Users only satisfy\n`user`. Creating, deleting, and ACL-updating a namespace are all\nadmin-only regardless of `write_role`.\n\n## No existence leak\nCallers who lack `read_role` receive 404 (not 403) on any operation\nagainst a namespace, so they cannot probe namespace existence.\n\n## Backups\nEvery successful mutation triggers an async R2 backup. Rapid bursts\ncoalesce into one backup per cooldown window (default 30s).\n",
    "title": "Config API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/config": {
      "get": {
        "description": "Returns all namespaces whose `read_role` is satisfied by the\ncaller. Admins see everything; users see only `read_role=user`\nnamespaces. Never returns 403 — invisible namespaces are simply\nomitted from the list.\n",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NamespaceSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Namespace summaries (may be an empty array)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "summary": "List namespaces visible to the caller",
        "tags": [
          "config"
        ]
      }
    },
    "/api/v1/config/namespaces": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNamespaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceInfo"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "summary": "Create a namespace (admin-only)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/config/namespaces/{ns}": {
      "parameters": [
        {
          "in": "path",
          "name": "ns",
          "required": true,
          "schema": {
            "pattern": "^[a-z0-9_-]{1,64}$",
            "type": "string"
          }
        }
      ],
      "patch": {
        "description": "Changes `read_role` and `write_role`; leaves the document untouched.\nThe same `write_role` ≥ `read_role` invariant applies as on create.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ACLUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceInfo"
                }
              }
            },
            "description": "Updated",
            "headers": {
              "X-Read-Role": {
                "description": "New read role after the update (`admin` or `user`)",
                "schema": {
                  "enum": [
                    "admin",
                    "user"
                  ],
                  "type": "string"
                }
              },
              "X-Write-Role": {
                "description": "New write role after the update (`admin` or `user`)",
                "schema": {
                  "enum": [
                    "admin",
                    "user"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "summary": "Update namespace ACL (admin-only)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/config/{ns}": {
      "delete": {
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "summary": "Delete a namespace (admin-only)",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceDocument"
                }
              }
            },
            "description": "Stored JSON document verbatim",
            "headers": {
              "Cache-Control": {
                "description": "Always `private, no-store`",
                "schema": {
                  "type": "string"
                }
              },
              "Vary": {
                "description": "Always `Authorization`",
                "schema": {
                  "type": "string"
                }
              },
              "X-Read-Role": {
                "description": "Role required to read this namespace (`admin` or `user`)",
                "schema": {
                  "enum": [
                    "admin",
                    "user"
                  ],
                  "type": "string"
                }
              },
              "X-Write-Role": {
                "description": "Role required to write this namespace (`admin` or `user`)",
                "schema": {
                  "enum": [
                    "admin",
                    "user"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "summary": "Fetch a namespace document",
        "tags": [
          "config"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "ns",
          "required": true,
          "schema": {
            "pattern": "^[a-z0-9_-]{1,64}$",
            "type": "string"
          }
        }
      ],
      "put": {
        "description": "Whole-document replace. Byte-identical content (after JSON\ncompaction) is detected server-side and returned as\n`changed: false` without a write or backup.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutResult"
                }
              }
            },
            "description": "Replaced (or confirmed no-op)"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "summary": "Replace a namespace document",
        "tags": [
          "config"
        ]
      }
    },
    "/healthz": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "example": "ok",
                      "type": "string"
                    },
                    "version": {
                      "example": "dev",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Service is up"
          }
        },
        "security": [],
        "summary": "Health probe",
        "tags": [
          "system"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://{domain}",
      "variables": {
        "domain": {
          "default": "config.example.com",
          "description": "Your config service hostname"
        }
      }
    }
  ],
  "tags": [
    {
      "description": "Namespace documents (read/write)",
      "name": "config"
    },
    {
      "description": "Namespace lifecycle (admin-only)",
      "name": "admin"
    },
    {
      "description": "Health and discovery",
      "name": "system"
    }
  ]
}